Fixed `RelatedInterestCategory`.
The ID given from the parliament API is not unique: categories 10a, 10b... all have the
same ID, and a different set of IDs is used per house of parliament. Ultimately, the
source API ID is useless and no longer used.

The sort order given by the parliament API is also unusable.

We now parse the category name to get major and minor 'codenames' and use those to
determine the correct sorting order (calculated after updates and persisted in the
`RelatedInterestCategory.sort_order` field).
e.g. "2. (b) Any other support not included in Category 2(a)"
     yields a codename of major=2, minor="b".