Improved resolution of `RegisteredInterest.description_data`.
Dates related to the administration of interest registration are now parsed
more aggressively from the description and stored separately from other fields.
There are often many of these dates, they can be ambiguous or contradictory,
and generally don't seem all that useful. We still keep track of them and make
them available, but doing this allows them to be hidden by default to reduce noise.