Improved map view.
National party territories are loaded as a background layer to provide color.
Constituency boundaries are layered on top showing only their outlines.
Inaccuracies in the simplified constituency outlines are not so obvious as
the gaps between boundaries do not leave gaps in the background color.

Constituencies are loaded using user location as the focus so nearby
constituencies load first. (Only if location permission is already allowed,
otherwise location of Parliament in London is used as the focus).