Added `--primary-container`/`--on-primary-container`.
This should be a variant of `--primary`/`--on-primary`, with adjusted
luminance/saturation to ensure it is suitable as a surface colour for elements
that may be large and/or contain text. In short, this colour should be
relatively understated.

These colours are now used for tooltips and `TintedButton`.

In contrast, `--primary` should be used for 'pop', sparingly applied as an
accent.